Completely overhaul loading logic
This allows us to have the scroller be ready before the render occurs
I had to pass in the initial values to the animationLogic because there’s a bug in reanimated where setting a value before render has no effect. Which means any call to “onUpdate” within InfiniteScroller won’t actually update the values in AnimationLogic. Going to maybe re-work that bit.